Training pipeline: metrics, early stopping, sliding window
- Add SlidingWindowDataset for efficient data utilization
- Implement EarlyStopping for robust training and reproducibility
- Add metrics.py: frame-level and note-level evaluation, velocity MAE
- Refactor ONNX export/verification to shared onnx_utils.py
- Make model/loss config more flexible via CLI args
- Improve error handling in MaestroDataset and throughout code
- Add set_seed() for reproducibility; CLI seed option
- Validation now computes and logs detailed metrics
- Add comprehensive pytest-based test_training.py
- Change C# PolyphonicTranscriptionResult.FrameRate to double
- Document collate_fn and improve code clarity
- Add alternative STFT (center padding) in MelSpectrogramExtractor (commented)
- General CLI, logging, and documentation improvements

+++ src/StaffSharp.MachineLearning/ML/Models/PolyphonicTranscriptionResult.cs
@@ -24,6 +24,7 @@ namespace StaffSharp.MachineLearning.ML.Models;
 /// <param name="FrameRate">
 /// Number of frames per second (Hz).
 /// Computed as SampleRate / HopSize.
+/// Uses double precision to avoid rounding errors (e.g., 16000 / 512 = 31.25 Hz).
 /// </param>
 /// <param name="SampleRate">
 /// Audio sample rate used for feature extraction (Hz).
@@ -33,7 +34,7 @@ public sealed record PolyphonicTranscriptionResult(
     float[,] OnsetRoll,
     float[,] OffsetRoll,
     float[,] VelocityRoll,
-    int FrameRate,
+    double FrameRate,
     int SampleRate)
 {
     /// <summary>
